Add files via upload
In this project, I accessed adataset and made necessary modifications to it through Python coding. The initial task involved correcting the prices in the dataset, which were incorrect. I achieved this by multiplying all price values by 0.9 to get the correct results.
After correcting the data, I created bar charts to visually represent both the original and adjusted prices. This allowed for a clear comparison between the incorrect and corrected values. The bar charts provided a comprehensive visualization of the data, making it easy to understand the impact of the correction.
The project demonstrated efficient data processing and visualization techniques using Python, ensuring that datasets could be manipulated and analyzed quickly and accurately.